backend/requirements/production.txt
-r base.txt

# Serveur
gunicorn>=22.0
psycopg2-binary>=2.9

# ML / IA
scikit-learn>=1.5
xgboost>=2.0
numpy>=1.26
pandas>=2.2
joblib>=1.4

# RAG / Embeddings
sentence-transformers>=3.0
faiss-cpu>=1.8
beautifulsoup4>=4.12
pdfplumber>=0.11
httpx>=0.27

# LLM clients
anthropic>=0.30
openai>=1.35

# Async / WebSocket
channels>=4.1
daphne>=4.1
celery>=5.4
redis>=5.0
